requirements.txt
amqp==5.2.0
asgiref==3.8.0
billiard==3.6.4.0
black==24.3.0
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.15.0
drf-yasg==1.21.7
flake8==7.0.0
idna==3.6
inflection==0.5.1
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
